sitemap.js: Inlined all helper functions => only one file.
